requirements.txt
langgraph==0.2.3
langchain==0.2.16
langchain-openai==0.1.23
python-dotenv==1.0.1
typing-extensions==4.12.2
slack-sdk==3.29.0
pandas==2.1.4
requests==2.31.0